Move WorkflowSpec to own file
