+++ tests/services/data_reduction_test.py
@@ -7,8 +7,9 @@ import numpy as np
 import pytest
 from streaming_data_types import eventdata_ev44
 
-from beamlime.config import models
+from beamlime.config import workflow_spec
 from beamlime.config.instruments import available_instruments
+from beamlime.config.models import ConfigKey, StartTime
 from beamlime.core.message import CONFIG_STREAM_ID
 from beamlime.handlers.config_handler import ConfigUpdate
 from beamlime.services.data_reduction import make_reduction_service_builder
@@ -16,8 +17,8 @@ from tests.helpers.beamlime_app import BeamlimeApp
 
 
 def _get_workflow_by_name(
-    workflow_specs: models.WorkflowSpecs, name: str
-) -> tuple[str, models.WorkflowSpec]:
+    workflow_specs: workflow_spec.WorkflowSpecs, name: str
+) -> tuple[str, workflow_spec.WorkflowSpec]:
     for wid, spec in workflow_specs.workflows.items():
         if spec.name == name:
             return wid, spec
@@ -38,8 +39,8 @@ def test_publishes_workflow_specs_on_startup(instrument: str) -> None:
     message = sink.messages[0]
     assert message.stream == CONFIG_STREAM_ID
     assert isinstance(message.value, ConfigUpdate)
-    assert isinstance(message.value.config_key, models.ConfigKey)
-    assert isinstance(message.value.value, models.WorkflowSpecs)
+    assert isinstance(message.value.config_key, ConfigKey)
+    assert isinstance(message.value.value, workflow_spec.WorkflowSpecs)
     if instrument in ('bifrost', 'dummy', 'loki'):
         assert len(message.value.value.workflows) > 0
 
@@ -58,10 +59,10 @@ def test_can_configure_and_stop_workflow_with_detector(
     sink.messages.clear()  # Clear the initial message
 
     # Assume workflow is runnable for all source names
-    config_key = models.ConfigKey(
+    config_key = ConfigKey(
         source_name=None, service_name="data_reduction", key="workflow_config"
     )
-    workflow_config = models.WorkflowConfig(
+    workflow_config = workflow_spec.WorkflowConfig(
         identifier=workflow_id,
         values={param.name: param.default for param in spec.parameters},
     )
@@ -113,10 +114,10 @@ def test_can_configure_and_stop_workflow_with_detector_and_monitors(
     sink.messages.clear()  # Clear the initial message
 
     # Assume workflow is runnable for all source names
-    config_key = models.ConfigKey(
+    config_key = ConfigKey(
         source_name=None, service_name="data_reduction", key="workflow_config"
     )
-    workflow_config = models.WorkflowConfig(
+    workflow_config = workflow_spec.WorkflowConfig(
         identifier=workflow_id,
         values={param.name: param.default for param in spec.parameters},
     )
@@ -176,10 +177,10 @@ def test_can_clear_workflow_via_config(caplog: pytest.LogCaptureFixture) -> None
     service.step()
 
     # Assume workflow is runnable for all source names
-    config_key = models.ConfigKey(
+    config_key = ConfigKey(
         source_name=None, service_name="data_reduction", key="workflow_config"
     )
-    workflow_config = models.WorkflowConfig(
+    workflow_config = workflow_spec.WorkflowConfig(
         identifier=workflow_id,
         values={param.name: param.default for param in spec.parameters},
     )
@@ -191,8 +192,8 @@ def test_can_clear_workflow_via_config(caplog: pytest.LogCaptureFixture) -> None
     assert len(sink.messages) == 2
     assert sink.messages[-1].value.values.sum() == 5000
 
-    config_key = models.ConfigKey(key="start_time")
-    model = models.StartTime(value=5, unit='s')
+    config_key = ConfigKey(key="start_time")
+    model = StartTime(value=5, unit='s')
     app.publish_config_message(key=config_key, value=model.model_dump())
 
     app.publish_events(size=1000, time=4)
@@ -202,8 +203,8 @@ def test_can_clear_workflow_via_config(caplog: pytest.LogCaptureFixture) -> None
     service.step()
     assert sink.messages[-1].value.values.sum() == 2000
 
-    config_key = models.ConfigKey(key="start_time")
-    model = models.StartTime(value=8, unit='s')
+    config_key = ConfigKey(key="start_time")
+    model = StartTime(value=8, unit='s')
     app.publish_config_message(key=config_key, value=model.model_dump())
 
     app.publish_events(size=100, time=9)
@@ -226,10 +227,10 @@ def test_service_can_recover_after_bad_workflow_id_was_set(
     sink.messages.clear()  # Clear the initial message
 
     # Assume workflow is runnable for all source names
-    config_key = models.ConfigKey(
+    config_key = ConfigKey(
         source_name=None, service_name="data_reduction", key="workflow_config"
     )
-    bad_workflow_id = models.WorkflowConfig(
+    bad_workflow_id = workflow_spec.WorkflowConfig(
         identifier='abcde12345',  # Invalid workflow ID
         values={param.name: param.default for param in spec.parameters},
     )
@@ -243,7 +244,7 @@ def test_service_can_recover_after_bad_workflow_id_was_set(
     service.step()
     assert len(sink.messages) == 0  # Workflow not started
 
-    bad_param_value = models.WorkflowConfig(
+    bad_param_value = workflow_spec.WorkflowConfig(
         identifier=workflow_id,
         values={param.name: param.default for param in spec.parameters},
     )
@@ -266,12 +267,14 @@ def test_service_can_recover_after_bad_workflow_param_was_set(
     sink.messages.clear()  # Clear the initial message
 
     # Assume workflow is runnable for all source names
-    config_key = models.ConfigKey(
+    config_key = ConfigKey(
         source_name=None, service_name="data_reduction", key="workflow_config"
     )
     defaults = {param.name: param.default for param in spec.parameters}
     defaults['does_not_exist'] = 1
-    bad_param_value = models.WorkflowConfig(identifier=workflow_id, values=defaults)
+    bad_param_value = workflow_spec.WorkflowConfig(
+        identifier=workflow_id, values=defaults
+    )
     # Trigger workflow start
     app.publish_config_message(key=config_key, value=bad_param_value.model_dump())
 
@@ -282,7 +285,7 @@ def test_service_can_recover_after_bad_workflow_param_was_set(
     service.step()
     assert len(sink.messages) == 0  # Workflow not started
 
-    bad_param_value = models.WorkflowConfig(
+    bad_param_value = workflow_spec.WorkflowConfig(
         identifier=workflow_id,
         values={param.name: param.default for param in spec.parameters},
     )
@@ -305,10 +308,10 @@ def test_active_workflow_keeps_running_when_bad_workflow_id_or_params_were_set(
     sink.messages.clear()  # Clear the initial message
 
     # Start a valid workflow first
-    config_key = models.ConfigKey(
+    config_key = ConfigKey(
         source_name=None, service_name="data_reduction", key="workflow_config"
     )
-    workflow_config = models.WorkflowConfig(
+    workflow_config = workflow_spec.WorkflowConfig(
         identifier=workflow_id,
         values={param.name: param.default for param in spec.parameters},
     )
@@ -322,7 +325,7 @@ def test_active_workflow_keeps_running_when_bad_workflow_id_or_params_were_set(
     assert sink.messages[0].value.values.sum() == 2000
 
     # Try to set an invalid workflow ID
-    bad_workflow_id = models.WorkflowConfig(
+    bad_workflow_id = workflow_spec.WorkflowConfig(
         identifier='abcde12345',  # Invalid workflow ID
         values={},
     )
@@ -337,7 +340,9 @@ def test_active_workflow_keeps_running_when_bad_workflow_id_or_params_were_set(
     # Try to set a workflow with invalid parameters
     defaults = {param.name: param.default for param in spec.parameters}
     defaults['does_not_exist'] = 1
-    bad_param_value = models.WorkflowConfig(identifier=workflow_id, values=defaults)
+    bad_param_value = workflow_spec.WorkflowConfig(
+        identifier=workflow_id, values=defaults
+    )
     app.publish_config_message(key=config_key, value=bad_param_value.model_dump())
 
     # Add more events and verify the original workflow is still running
@@ -376,10 +381,10 @@ def test_workflow_starts_with_specific_or_global_source_name(
         service.step()
         assert len(sink.messages) == 0
 
-    config_key = models.ConfigKey(
+    config_key = ConfigKey(
         source_name=source_name, service_name="data_reduction", key="workflow_config"
     )
-    workflow_config = models.WorkflowConfig(
+    workflow_config = workflow_spec.WorkflowConfig(
         identifier=workflow_id,
         values={param.name: param.default for param in spec.parameters},
     )
@@ -405,10 +410,10 @@ def configured_dummy_reduction() -> BeamlimeApp:
     workflow_id, spec = _get_workflow_by_name(workflow_specs, 'Total counts')
     sink.messages.clear()  # Clear the initial message
 
-    config_key = models.ConfigKey(
+    config_key = ConfigKey(
         source_name='panel_0', service_name="data_reduction", key="workflow_config"
     )
-    workflow_config = models.WorkflowConfig(
+    workflow_config = workflow_spec.WorkflowConfig(
         identifier=workflow_id,
         values={param.name: param.default for param in spec.parameters},
     )
